Add simplificiation logic
