+++ dashie/src/md_doc/hir/mod.rs
@@ -1,207 +1,264 @@
-mod doc;
+mod node;
 
-pub(crate) use doc::*;
+pub(crate) use node::*;
 
-use crate::dashie_schema;
+mod simplifying;
+
 use crate::prelude::*;
-use serde_json::Value;
-use std::fmt;
+use crate::source::{ArraySchema, ObjectSchema, OneOfVariantSchema, RootSchema};
+use buildstructor::buildstructor;
+use std::collections::BTreeMap;
 
-#[derive(Debug)]
-pub(crate) struct SchemaNode {
-    pub(crate) schema: Schema,
-    pub(crate) children: Vec<SchemaNode>,
+pub(crate) struct Dom {
+    pub(crate) root: SchemaNode,
+    pub(crate) type_index: BTreeMap<String, SchemaNode>,
 }
 
-impl SchemaNode {
-    pub(crate) fn leaf(schema: Schema) -> Self {
-        Self {
-            schema,
-            children: vec![],
-        }
-    }
-}
+#[buildstructor]
+impl Dom {
+    #[builder]
+    pub(crate) fn new(
+        schema: RootSchema,
+        max_nesting_in_file: Option<u8>,
+        allow_unused_definitions: Option<bool>,
+    ) -> Result<Self> {
+        let ctx = LoweringContext {
+            root: schema,
+            max_nesting_in_file: max_nesting_in_file.unwrap_or(2),
+            allow_unused_definitions: allow_unused_definitions.unwrap_or(false),
+        };
 
-#[derive(Debug)]
-pub(crate) struct Schema {
-    pub(crate) path: Path,
-    pub(crate) ty: Type,
-    pub(crate) doc: SchemaDoc,
-}
+        let doc = ctx.lower()?;
 
-#[derive(Debug)]
-pub(crate) enum SchemaDoc {
-    /// Documentation should be embedded in the same file. The value is [`None`]
-    Embedded(SchemaDocData),
-
-    /// Schema should live as a nested document
-    Nested(SchemaDocData),
-
-    /// This schema is a reference to some other schema. It may be either a reference
-    /// to a definition within the same schema or a reference to some external schema.
-    Ref(SchemaDocRef),
-}
-
-impl SchemaDoc {
-    pub(crate) fn reference(&self) -> Option<&str> {
-        match self {
-            SchemaDoc::Ref(reference) => Some(&reference.reference),
-            _ => None,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct SchemaDocRef {
-    pub(crate) reference: String,
-
-    /// Additional data that may override the details of the referenced schema.
-    pub(crate) data: SchemaDocData,
-}
-
-#[derive(Debug)]
-pub(crate) struct SchemaDocData {
-    pub(crate) title: Option<String>,
-    pub(crate) description: Option<String>,
-    pub(crate) default: Option<Value>,
-    pub(crate) examples: Vec<Value>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct Path {
-    pub(crate) origin: PathOrigin,
-    pub(crate) segments: Vec<PathSegment>,
-}
-
-impl Path {
-    pub(crate) fn new(origin: PathOrigin) -> Self {
-        Self {
-            origin,
-            segments: vec![],
-        }
+        Ok(doc)
     }
 
-    #[must_use]
-    pub(crate) fn add_segment(&self, new_segment: PathSegment) -> Self {
-        let mut segments = self.segments.clone();
-        segments.push(new_segment);
-        Self {
-            origin: self.origin.clone(),
-            segments,
-        }
-    }
+    // fn flatten(&mut self, max_level: usize) -> Result<()> {
+    //     FlattenDoc {
+    //         ty_index: &mut self.type_index,
+    //         max_level,
+    //     }
+    //     .flatten(&mut self.root, 0)
+    // }
 }
 
-impl fmt::Display for Path {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut segments = self.segments.iter();
+// struct FlattenDoc<'a> {
+//     ty_index: &'a mut BTreeMap<String, TypeDocNode>,
+//     max_level: usize,
+// }
 
-        if let Some(segment) = segments.next() {
-            if let PathSegment::Variant(_) = segment {
-                write!(f, "As {segment}")?;
-            } else {
-                write!(f, "{segment}")?;
-            }
-        }
+// impl FlattenDoc<'_> {
+//     fn flatten(&mut self, node: &mut TypeDocNode, level: usize) -> Result<()> {
+//         for child in &mut node.children {
+//             if level <= self.max_level {
+//                 Self {
+//                     ty_index: self.ty_index,
+//                     max_level: self.max_level,
+//                 }
+//             }
 
-        segments.try_for_each(|segment| match &segment {
-            PathSegment::Field(_) => write!(f, ".{segment}"),
-            PathSegment::Index => write!(f, "{segment}"),
-            PathSegment::Variant(_) => write!(f, "as {segment}"),
+//             self.flatten(child, level + 1)?;
+//         }
+
+//         if level <= self.max_level {
+//             return Ok(());
+//         }
+
+//         node.children.clear();
+
+//         let type_doc = &node.inner;
+
+//         let definition = type_doc
+//             .key
+//             .definition()
+//             .with_context(|| {
+//                 format!(
+//                     "Can't flatten node at level {level}, because the name to \
+//                     assign to it in the type index can not be inferred.\n\
+//                     Schema key: {}\n\
+//                     Try moving the schema to a definition, and the definition key \
+//                     will be used as a name for this type in the type index",
+//                     type_doc.key
+//                 )
+//             })?
+//             .to_owned();
+
+//         let type_index_entry = self.ty_index.get(&definition).unwrap_or_else(|| {
+//             panic!("We inlined this type before, so it must be in type index: {definition}")
+//         });
+
+//         let type_index_ref = TypeIndexRef {
+//             definition: definition.clone(),
+//             ty: type_index_entry.inner.ty.clone(),
+//         };
+
+//         let new_type_doc = TypeDoc {
+//             key: type_doc.key.clone(),
+//             title: None,
+//             description: None,
+//             default: None,
+//             examples: vec![],
+//             ty: type_doc.ty.clone(),
+//             type_index_ref: Some(type_index_ref),
+//         };
+
+//         let node = std::mem::replace(node, TypeDocNode::leaf(new_type_doc));
+
+//         self.ty_index.insert(definition, node);
+
+//         Ok(())
+//     }
+// }
+
+struct LoweringContext {
+    root: RootSchema,
+    max_nesting_in_file: u8,
+    allow_unused_definitions: bool,
+}
+
+impl LoweringContext {
+    fn lower(mut self) -> Result<Dom> {
+        self.simplify()?;
+
+        let root = PathedSourceSchema::origin(PathOrigin::Root, self.root.schema.clone());
+        let root = self.schema_node(root)?;
+
+        let definitions = self
+            .root
+            .definitions
+            .iter()
+            .map(|(def_name, schema)| {
+                let origin = PathOrigin::Definition(def_name.clone());
+                let schema = PathedSourceSchema::origin(origin, schema.clone());
+
+                Ok((def_name.clone(), self.schema_node(schema)?))
+            })
+            .collect::<Result<_>>()?;
+
+        Ok(Dom {
+            root,
+            type_index: definitions,
         })
     }
-}
 
-#[derive(Clone, Debug)]
-pub(crate) enum PathOrigin {
-    Root,
-    Definition(String),
-}
+    fn schema_node(&self, schema: PathedSourceSchema) -> Result<SchemaNode> {
+        //     if let Some(reference) = schema.inner.reference.clone() {
+        //         self.schema_node_ref(schema, reference)
+        //     } else {
+        //         self.schema_node_non_ref(schema)
+        //     }
+        // }
 
-impl fmt::Display for PathSegment {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self {
-            PathSegment::Field(field) => f.write_str(&field.name),
-            // .<nth>
-            // [..]
-            PathSegment::Index => f.write_str("[N]"),
-            PathSegment::Variant(name) => write!(f, "{name}"),
-        }
-    }
-}
+        // fn schema_node_ref(&self, schema: PathedSourceSchema, reference: String) -> Result<SchemaNode> {
+        //     let doc = SchemaDoc::Ref(reference);
+        //     let path = schema.path;
 
-#[derive(Clone, Debug)]
-pub(crate) enum PathSegment {
-    Field(Field),
-    Index,
-    Variant(String),
-}
+        //     let inline = self
+        //         .options
+        //         .root
+        //         .inline_referenced_definition(&schema.inner)?;
 
-#[derive(Clone, Debug)]
-pub(crate) struct Field {
-    pub(crate) name: String,
-    pub(crate) required: bool,
-}
+        //     let ty = Type::from_source_schema(&inline);
 
-#[derive(Debug, Clone)]
-pub(crate) struct LeafType {
-    pub(crate) ty: Option<String>,
-    pub(crate) format: Option<String>,
-    pub(crate) enum_schema: Option<Vec<dashie_schema::EnumVariantSchema>>,
-}
+        //     let schema = Schema { path, ty, doc };
 
-impl LeafType {
-    fn from_dashie_schema(schema: &dashie_schema::Schema) -> Self {
-        Self {
-            ty: schema.ty.clone(),
-            format: schema.format.clone(),
-            enum_schema: schema.enum_schema.clone(),
-        }
-    }
-}
+        //     Ok(SchemaNode::leaf(schema))
+        // }
 
-#[derive(Debug, Clone)]
-pub(crate) struct Type {
-    pub(crate) inner: LeafType,
+        // fn schema_node_non_ref(&self, schema: PathedSourceSchema) -> Result<SchemaNode> {
+        let path = schema.path;
 
-    /// [`LeafType`] exists to make sure we don't have recursion in this field.
-    /// We describe at max two levels of nesting in type tags.
-    pub(crate) array_items_ty: Option<LeafType>,
-}
+        let ty = Type::from_source_schema(&self.root.inline_referenced_definition(&schema.inner)?);
 
-impl Type {
-    pub(crate) fn from_source_schema(schema: &dashie_schema::Schema) -> Self {
-        let inner = LeafType::from_dashie_schema(schema);
+        let schema = schema.inner;
 
-        let array_items_ty = schema
-            .array_schema
-            .as_ref()
-            .map(|array| LeafType::from_dashie_schema(&array.items));
+        let children = if let Some(array) = schema.array_schema {
+            Self::array_children(&path, array)?
+        } else if let Some(object) = schema.object_schema {
+            Self::object_children(&path, object)?
+        } else if let Some(variants) = schema.one_of {
+            Self::one_of_children(&path, variants)?
+        } else {
+            vec![]
+        };
 
-        Self {
-            inner,
-            array_items_ty,
-        }
-    }
-}
+        let children = children
+            .into_iter()
+            .map(|child| self.schema_node(child))
+            .try_collect()?;
 
-#[derive(Clone, Debug)]
-pub(crate) struct PathedSourceSchema {
-    pub(crate) path: Path,
-    pub(crate) inner: dashie_schema::Schema,
-}
+        let data = SchemaDocData {
+            title: schema.title,
+            description: schema.description,
+            default: schema.default,
+            examples: schema.examples,
+        };
 
-impl PathedSourceSchema {
-    pub(crate) fn new(path: Path, inner: dashie_schema::Schema) -> Self {
-        Self { path, inner }
+        let doc = if let Some(reference) = schema.reference.clone() {
+            SchemaDoc::Ref(SchemaDocRef { reference, data })
+        } else if path.segments.len() % (usize::from(self.max_nesting_in_file) + 1) == 0 {
+            SchemaDoc::Nested(data)
+        } else {
+            SchemaDoc::Embedded(data)
+        };
+
+        let schema = Schema { path, ty, doc };
+
+        Ok(SchemaNode { schema, children })
     }
 
-    pub(crate) fn origin(origin: PathOrigin, inner: dashie_schema::Schema) -> Self {
-        Self::new(Path::new(origin), inner)
+    fn array_children(path: &Path, array: ArraySchema) -> Result<Vec<PathedSourceSchema>> {
+        let path = path.add_segment(PathSegment::Index);
+        let items = PathedSourceSchema::new(path, *array.items);
+        Ok(vec![items])
     }
 
-    fn inline_referenced_definition(mut self, root: &dashie_schema::RootSchema) -> Result<Self> {
-        self.inner = root.inline_referenced_definition(&self.inner)?;
-        Ok(self)
+    fn object_children(path: &Path, object: ObjectSchema) -> Result<Vec<PathedSourceSchema>> {
+        let properties = object
+            .properties
+            .into_iter()
+            .map(|(name, value)| {
+                let field = Field {
+                    name: name.clone(),
+                    required: object.required.contains(&name),
+                };
+                let path = path.add_segment(PathSegment::Field(field));
+                PathedSourceSchema::new(path, value)
+            })
+            .collect();
+
+        Ok(properties)
+    }
+
+    fn one_of_children(
+        path: &Path,
+        variants: Vec<OneOfVariantSchema>,
+    ) -> Result<Vec<PathedSourceSchema>> {
+        let names: Vec<_> = variants
+            .clone()
+            .into_iter()
+            .map(|variant| variant.name().map(ToOwned::to_owned))
+            .try_collect()?;
+
+        let duplicates: Vec<_> = names.iter().duplicates().collect();
+
+        anyhow::ensure!(
+            duplicates.is_empty(),
+            "Duplicate variant names found in one_of schema.\n\
+            Duplicates: {duplicates:?}\n\
+            Variants: {variants:#?}",
+        );
+
+        let variants = variants
+            .into_iter()
+            .zip(names)
+            .map(|(variant, name)| {
+                let path = path.add_segment(PathSegment::Variant(name));
+
+                PathedSourceSchema::new(path, variant.schema)
+            })
+            .collect();
+
+        Ok(variants)
     }
 }
